#72 Implement path from search bar
The map automatically adds a path from the GPS location to the location that was searched. The code was cleaned as well to be more modulable and to be able to add points on different actions, in this case, both when you search for a location and when you click on the map.
