+++ arduinoRemote/app/src/main/java/com/example/arduinoremote/FetchUrl.java
@@ -1,0 +1,89 @@
+package com.example.arduinoremote;
+
+import android.os.AsyncTask;
+import android.util.Log;
+
+import com.google.android.gms.maps.GoogleMap;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.URL;
+
+class FetchUrl extends AsyncTask<String, Void, String> {
+
+    private GoogleMap map;
+
+    public FetchUrl(GoogleMap map){
+        this.map = map;
+    }
+
+
+    @Override
+    protected String doInBackground(String... url) {
+
+        // For storing data from web service
+        String data = "";
+
+        try {
+            // Fetching the data from web service
+            data = downloadUrl(url[0]);
+            Log.d("Background Task data", data.toString());
+        } catch (Exception e) {
+            Log.d("Background Task", e.toString());
+        }
+        return data;
+    }
+
+    @Override
+    protected void onPostExecute(String result) {
+        super.onPostExecute(result);
+
+        ParserTask parserTask = new ParserTask(map);
+
+        // Invokes the thread for parsing the JSON data
+        parserTask.execute(result);
+
+
+    }
+
+    private String downloadUrl(String strUrl) throws IOException {
+        String data = "";
+        InputStream iStream = null;
+        HttpURLConnection urlConnection = null;
+        try {
+            URL url = new URL(strUrl);
+
+            // Creating an http connection to communicate with url
+            urlConnection = (HttpURLConnection) url.openConnection();
+
+            // Connecting to url
+            urlConnection.connect();
+
+            // Reading data from url
+            iStream = urlConnection.getInputStream();
+
+            BufferedReader br = new BufferedReader(new InputStreamReader(iStream));
+
+            StringBuffer sb = new StringBuffer();
+
+            String line = "";
+            while ((line = br.readLine()) != null) {
+                sb.append(line);
+            }
+
+            data = sb.toString();
+            Log.d("downloadUrl", data.toString());
+            br.close();
+
+        } catch (Exception e) {
+            Log.d("Exception", e.toString());
+        } finally {
+            iStream.close();
+            urlConnection.disconnect();
+        }
+        return data;
+    }
+}
