+++ arduinoRemote/app/src/main/java/com/example/arduinoremote/MapsActivity.java
@@ -53,6 +53,36 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
         SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
                 .findFragmentById(R.id.map);
         mapFragment.getMapAsync(this);
+        
+    }
+
+    /**
+     * Manipulates the map once available.
+     * This callback is triggered when the map is ready to be used.
+     * This is where we can add markers or lines, add listeners or move the camera. In this case,
+     * we just add a marker near Sydney, Australia.
+     * If Google Play services is not installed on the device, the user will be prompted to install
+     * it inside the SupportMapFragment. This method will only be triggered once the user has
+     * installed Google Play services and returned to the app.
+     */
+
+    @Override
+    public void onMapReady(GoogleMap googleMap) {
+        mMap = googleMap;
+        final ArrayList<LatLng> MarkerPoints = new ArrayList<>();
+
+        // Marker is set at Car position and the camera is there as well
+         LatLng carPosition = new LatLng((ConnectBT.getCoordinates()[0]), ConnectBT.getCoordinates()[1] );
+        mMap.addMarker(new MarkerOptions().position(carPosition).title("Marker at Car").icon((BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_GREEN))));
+        mMap.moveCamera(CameraUpdateFactory.newLatLng(carPosition));
+        mMap.setOnMapClickListener(new GoogleMap.OnMapClickListener() {
+
+            @Override
+            public void onMapClick(LatLng point) {
+addPoint(MarkerPoints, point);
+addPath(MarkerPoints);
+            }
+        });
 
         /**
          * Initialize Places. For simplicity, the API key is hard-coded. In a production
@@ -73,9 +103,8 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
             @Override
             public void onPlaceSelected(Place place) {
                 if (place.getLatLng() != null){
-                    mMap.addMarker(new MarkerOptions().position(place.getLatLng()).icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_RED)));
-                    mMap.moveCamera(CameraUpdateFactory.newLatLng(place.getLatLng()));
-                    mMap.animateCamera(CameraUpdateFactory.zoomTo(15));
+                    addPoint(MarkerPoints, place.getLatLng());
+                    addPath(MarkerPoints);
                 }
                 Log.i("PlaceLog", "Place: " + place.getName() + ", " + place.getId() + "," +place.getLatLng()) ;
             }
@@ -88,332 +117,67 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
         });
     }
 
-
-    /**
-     * Manipulates the map once available.
-     * This callback is triggered when the map is ready to be used.
-     * This is where we can add markers or lines, add listeners or move the camera. In this case,
-     * we just add a marker near Sydney, Australia.
-     * If Google Play services is not installed on the device, the user will be prompted to install
-     * it inside the SupportMapFragment. This method will only be triggered once the user has
-     * installed Google Play services and returned to the app.
-     */
-
-
-
-    @Override
-    public void onMapReady(GoogleMap googleMap) {
-        mMap = googleMap;
-        final ArrayList<LatLng> MarkerPoints = new ArrayList<>();
-
-        // Marker is set at Car position and the camera is there as well
-         LatLng carPosition = new LatLng((ConnectBT.getCoordinates()[0]), ConnectBT.getCoordinates()[1] );
+    public void addPoint(ArrayList<LatLng> MarkerPoints, LatLng point){
+        LatLng carPosition = new LatLng((ConnectBT.getCoordinates()[0]), ConnectBT.getCoordinates()[1] );
         mMap.addMarker(new MarkerOptions().position(carPosition).title("Marker at Car").icon((BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_GREEN))));
-        mMap.moveCamera(CameraUpdateFactory.newLatLng(carPosition));
-        mMap.setOnMapClickListener(new GoogleMap.OnMapClickListener() {
 
-            @Override
-            public void onMapClick(LatLng point) {
-
-                 LatLng carPosition = new LatLng((ConnectBT.getCoordinates()[0]), ConnectBT.getCoordinates()[1] );
-                mMap.addMarker(new MarkerOptions().position(carPosition).title("Marker at Car").icon((BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_GREEN))));
-
-                // Already one location
-                if (MarkerPoints.size() > 0) {
-                    MarkerPoints.clear();
-                    mMap.clear();
-                    mMap.addMarker(new MarkerOptions().position(carPosition).title("Marker at Car").icon((BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_GREEN))));
-                }
-
-                // Adding new item to the ArrayList
-                MarkerPoints.add(point);
-
-                // Creating MarkerOptions
-                MarkerOptions options = new MarkerOptions();
-
-                // Setting the position of the marker
-                options.position(point);
-
-                /**
-                 * For the start location, the color of marker is GREEN and
-                 * for the end location, the color of marker is RED.
-                 */
-                if (MarkerPoints.size() == 1) {
-                    options.icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_RED));
-
-                }
-
-
-                // Add new marker to the Google Map Android API V2
-                mMap.addMarker(options);
-
-                // Checks, whether start and end locations are captured
-                if (MarkerPoints.size() >= 1) {
-                    LatLng origin = carPosition;
-                    LatLng dest = MarkerPoints.get(0);
-                    /*Polyline polyline1 = mMap.addPolyline(new PolylineOptions()
-                            .clickable(true)
-                            .add(
-                                    origin,
-                                    dest));*/
-
-
-
-                    // Getting URL to the Google Directions API
-                    String url = getUrl(origin, dest);
-                    Log.d("onMapClick", url);
-                    FetchUrl FetchUrl = new FetchUrl();
-
-                    // Start downloading json data from Google Directions API
-                    FetchUrl.execute(url);
-
-
-                    //move map camera
-                    mMap.moveCamera(CameraUpdateFactory.newLatLng(origin));
-                    mMap.animateCamera(CameraUpdateFactory.zoomTo(15));
-                }
-
-            }
-
-
-
-            private String getUrl(LatLng origin, LatLng dest) {
-                return "https://maps.googleapis.com/maps/api/directions/json?" +
-                        "origin=" + origin.latitude + ","+ origin.longitude + "&destination=" + dest.latitude +","+ dest.longitude +
-                        "&key="+ getResources().getString(R.string.google_maps_key);
-            }
-        });
-    }
-
-    class ParserTask extends AsyncTask<String, Integer, List<List<HashMap<String, String>>>> {
-
-        // Parsing the data in non-ui thread
-        @Override
-        protected List<List<HashMap<String, String>>> doInBackground(String... jsonData) {
-
-            JSONObject jObject;
-            List<List<HashMap<String, String>>> routes = null;
-
-            try {
-                jObject = new JSONObject(jsonData[0]);
-                Log.d("ParserTask", jsonData[0].toString());
-                DataParser parser = new DataParser();
-                Log.d("ParserTask", parser.toString());
-
-                // Starts parsing data
-                routes = parser.parse(jObject);
-                Log.d("ParserTask", "Executing routes");
-                Log.d("ParserTask", routes.toString());
-
-            } catch (Exception e) {
-                Log.d("ParserTask", e.toString());
-                e.printStackTrace();
-            }
-            return routes;
+        // Already one location
+        if (MarkerPoints.size() > 0) {
+            MarkerPoints.clear();
+            mMap.clear();
+            mMap.addMarker(new MarkerOptions().position(carPosition).title("Marker at Car").icon((BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_GREEN))));
         }
 
-        // Executes in UI thread, after the parsing process
-        @Override
-        protected void onPostExecute(List<List<HashMap<String, String>>> result) {
-            ArrayList<LatLng> points;
-            PolylineOptions lineOptions = null;
+        // Adding new item to the ArrayList
+        MarkerPoints.add(point);
 
-            // Traversing through all the routes
-            for (int i = 0; i < result.size(); i++) {
-                points = new ArrayList<>();
-                lineOptions = new PolylineOptions();
+        // Creating MarkerOptions
+        MarkerOptions options = new MarkerOptions();
 
-                // Fetching i-th route
-                List<HashMap<String, String>> path = result.get(i);
-
-                // Fetching all the points in i-th route
-                for (int j = 0; j < path.size(); j++) {
-                    HashMap<String, String> point = path.get(j);
-
-                    double lat = Double.parseDouble(point.get("lat"));
-                    double lng = Double.parseDouble(point.get("lng"));
-                    LatLng position = new LatLng(lat, lng);
-
-                    points.add(position);
-                }
-
-                // Adding all the points in the route to LineOptions
-                lineOptions.addAll(points);
-                lineOptions.width(10);
-                lineOptions.color(Color.RED);
-
-                Log.d("onPostExecute", "onPostExecute lineoptions decoded");
-
-            }
-
-            // Drawing polyline in the Google Map for the i-th route
-            if (lineOptions != null) {
-
-                mMap.addPolyline(lineOptions);
-            } else {
-                Log.d("onPostExecute", "without Polylines drawn");
-            }
-        }
-
-    }
-
-    class FetchUrl extends AsyncTask<String, Void, String> {
-
-        @Override
-        protected String doInBackground(String... url) {
-
-            // For storing data from web service
-            String data = "";
-
-            try {
-                // Fetching the data from web service
-                data = downloadUrl(url[0]);
-                Log.d("Background Task data", data.toString());
-            } catch (Exception e) {
-                Log.d("Background Task", e.toString());
-            }
-            return data;
-        }
-
-        @Override
-        protected void onPostExecute(String result) {
-            super.onPostExecute(result);
-
-            ParserTask parserTask = new ParserTask();
-
-            // Invokes the thread for parsing the JSON data
-            parserTask.execute(result);
-
-
-        }
-
-        private String downloadUrl(String strUrl) throws IOException {
-            String data = "";
-            InputStream iStream = null;
-            HttpURLConnection urlConnection = null;
-            try {
-                URL url = new URL(strUrl);
-
-                // Creating an http connection to communicate with url
-                urlConnection = (HttpURLConnection) url.openConnection();
-
-                // Connecting to url
-                urlConnection.connect();
-
-                // Reading data from url
-                iStream = urlConnection.getInputStream();
-
-                BufferedReader br = new BufferedReader(new InputStreamReader(iStream));
-
-                StringBuffer sb = new StringBuffer();
-
-                String line = "";
-                while ((line = br.readLine()) != null) {
-                    sb.append(line);
-                }
-
-                data = sb.toString();
-                Log.d("downloadUrl", data.toString());
-                br.close();
-
-            } catch (Exception e) {
-                Log.d("Exception", e.toString());
-            } finally {
-                iStream.close();
-                urlConnection.disconnect();
-            }
-            return data;
-        }
-    }
-
-    class DataParser {
-
-        List<List<HashMap<String, String>>> parse(JSONObject jObject) {
-
-            List<List<HashMap<String, String>>> routes = new ArrayList<>();
-            JSONArray jRoutes;
-            JSONArray jLegs;
-            JSONArray jSteps;
-
-            try {
-
-                jRoutes = jObject.getJSONArray("routes");
-
-                /** Traversing all routes */
-                for (int i = 0; i < jRoutes.length(); i++) {
-                    jLegs = ((JSONObject) jRoutes.get(i)).getJSONArray("legs");
-                    List path = new ArrayList<>();
-
-                    /** Traversing all legs */
-                    for (int j = 0; j < jLegs.length(); j++) {
-                        jSteps = ((JSONObject) jLegs.get(j)).getJSONArray("steps");
-
-                        /** Traversing all steps */
-                        for (int k = 0; k < jSteps.length(); k++) {
-                            String polyline = "";
-                            polyline = (String) ((JSONObject) ((JSONObject) jSteps.get(k)).get("polyline")).get("points");
-                            List<LatLng> list = decodePoly(polyline);
-
-                            /** Traversing all points */
-                            for (int l = 0; l < list.size(); l++) {
-                                HashMap<String, String> hm = new HashMap<>();
-                                hm.put("lat", Double.toString((list.get(l)).latitude));
-                                hm.put("lng", Double.toString((list.get(l)).longitude));
-                                path.add(hm);
-                            }
-                        }
-                        routes.add(path);
-                    }
-                }
-
-            } catch (JSONException e) {
-                e.printStackTrace();
-            } catch (Exception e) {
-            }
-
-
-            return routes;
-        }
-
+        // Setting the position of the marker
+        options.position(point);
 
         /**
-         * Method to decode polyline points
+         * For the start location, the color of marker is GREEN and
+         * for the end location, the color of marker is RED.
          */
-        private List<LatLng> decodePoly(String encoded) {
+        if (MarkerPoints.size() == 1) {
+            options.icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_RED));
 
-            List<LatLng> poly = new ArrayList<>();
-            int index = 0, len = encoded.length();
-            int lat = 0, lng = 0;
-
-            while (index < len) {
-                int b, shift = 0, result = 0;
-                do {
-                    b = encoded.charAt(index++) - 63;
-                    result |= (b & 0x1f) << shift;
-                    shift += 5;
-                } while (b >= 0x20);
-                int dlat = ((result & 1) != 0 ? ~(result >> 1) : (result >> 1));
-                lat += dlat;
-
-                shift = 0;
-                result = 0;
-                do {
-                    b = encoded.charAt(index++) - 63;
-                    result |= (b & 0x1f) << shift;
-                    shift += 5;
-                } while (b >= 0x20);
-                int dlng = ((result & 1) != 0 ? ~(result >> 1) : (result >> 1));
-                lng += dlng;
-
-                LatLng p = new LatLng((((double) lat / 1E5)),
-                        (((double) lng / 1E5)));
-                poly.add(p);
-            }
-
-            return poly;
         }
 
 
+        // Add new marker to the Google Map Android API V2
+        mMap.addMarker(options);
+
+
+    }
+
+    public void addPath(ArrayList<LatLng> MarkerPoints){
+        // Checks, whether start and end locations are captured
+        if (MarkerPoints.size() >= 1) {
+            LatLng origin = new LatLng((ConnectBT.getCoordinates()[0]), ConnectBT.getCoordinates()[1] );
+            LatLng dest = MarkerPoints.get(0);
+
+            //Prepare request to Google Directions API
+            String url = getUrl(origin, dest);
+            Log.d("onMapClick", url);
+            FetchUrl FetchUrl = new FetchUrl(mMap);
+
+            // Start downloading json data from Google Directions API
+            FetchUrl.execute(url);
+
+
+            //move map camera
+            mMap.moveCamera(CameraUpdateFactory.newLatLng(origin));
+            mMap.animateCamera(CameraUpdateFactory.zoomTo(15));
+        }
+    }
+
+    private String getUrl(LatLng origin, LatLng dest) {
+        return "https://maps.googleapis.com/maps/api/directions/json?" +
+                "origin=" + origin.latitude + ","+ origin.longitude + "&destination=" + dest.latitude +","+ dest.longitude +
+                "&key="+ getResources().getString(R.string.google_maps_key);
     }
 }
